+++ Codingteam.Site/Clock.fs
@@ -1,7 +1,0 @@
-﻿namespace Codingteam.Site
-
-open System
-
-type Clock(getter: unit -> DateTime) =
-    member _.GetUtcDateTime(): DateTime = getter()
-    static member Default: Clock = Clock(fun () -> DateTime.UtcNow)