+++ Codingteam.Site/Controllers/HomeController.fs
@@ -1,29 +1,10 @@
 namespace Codingteam.Site.Controllers
 
-open System
-open System.Globalization
-
 open Microsoft.AspNetCore.Mvc
-open Microsoft.Extensions.Options
 
-open Codingteam.Site
-
-type HomeController(options: IOptions<CtorSettings>, clock: Clock) =
+type HomeController() =
     inherit Controller()
 
-    let config = options.Value
-    let getLogUrl (date : DateTime) =
-        let dateString = date.ToString("yyyy/MM/dd", CultureInfo.InvariantCulture)
-        sprintf "%s/%s.html" config.LogUrlPrefix dateString
-    let logTimezone = TimeSpan.FromHours <| double config.LogTimeZoneOffset
-    let getTodayLogUrl () = getLogUrl <| clock.GetUtcDateTime() + logTimezone
-
     member this.Index() =
-        this.ViewData.["LogUrl"] <- getTodayLogUrl ()
-        this.ViewData.["Title"] <- "Logs — Codingteam"
-        this.View()
-
-    [<Route("resources")>]
-    member this.Resources() =
-        this.ViewData.["Title"] <- "Resources — Codingteam"
+        this.ViewData.["Title"] <- "Codingteam"
         this.View()