+++ Codingteam.Site/CtorSettings.fs
@@ -1,5 +1,0 @@
-namespace Codingteam.Site
-
-type CtorSettings() =
-    member val LogUrlPrefix = "" with get, set
-    member val LogTimeZoneOffset = 0 with get, set